DefaultHttpDataSource: Fix can not play through a local proxy
When playing the video which is a local address eg: http://127.0.0.1 or htttp://10.66.66.66. Such address can be served by NanoHTTPD or something similar.

It’s commonly in a enterprise environment that everything is behind a proxy. when user have a proxy setting on the phone. The playback will fail because the proxy server can not connect to a local adress such as 127.0.0.1

So, we should bypass the local proxy setting in such scenario.