Remove user agent override from OkHttpDataSourceFactory
Like any other header, the `User-Agent` can be configured on an `OkHttpClient`
instance e.g.:

```kotlin
val okHttpClient = OkHttpClient.Builder()
        .addInterceptor { chain ->
            val request = chain.request().newBuilder().addHeader(USER_AGENT, userAgent).build()
            chain.proceed(request)
        }
        .addInterceptor(httpLoggingInterceptor)
        .build()
```

The `OkHttpDataSource` required that a `User-Agent` was passed in as a parameter
overriding any `User-Agent` already configured on the `OkHttpClient`. This made
it impossible to configure the `User-Agent` in once place if the `OkHttpClient`
is also used elsewhere in the app.

The OkHttpClient is the appropriate place to configure headers that are not
specific to a particular request.